requirements.txt
librosa
pytube
ffmpeg-python
pathvalidate